requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
